requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
itsdangerous
Werkzeug
